+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Show version (@ha62791)
 - Show current user (@ha62791)
 - List tables (@ha62791)
+- Check if table exists (@ha62791)